+++ CHANGELOG.md
@@ -80,6 +80,8 @@ and this project adheres to
   queriers directly. ([#1977])
 - cosmwasm-vm: Rename `BackendApi::canonical_address`/`::human_address` to
   `::addr_canonicalize`/`::addr_humanize` for consistency.
+- cosmwasm-vm: Add `BackendApi::addr_validate` to avoid having to do two calls
+  from Rust into Go.
 
 [#1874]: https://github.com/CosmWasm/cosmwasm/pull/1874
 [#1876]: https://github.com/CosmWasm/cosmwasm/pull/1876